Factored out the options class into a separate header
